multi: Update macdrive -> Darwinkit (#3264)
* upgrade to darwinkit


* bisonw-desktop: fix conversion


---------

+++ client/cmd/bisonw-desktop/app.go
@@ -430,9 +430,9 @@ func runWebview(url string) {
 		useIconBytes(w, Icon) // useIcon(w, "src/dexc.png")
 	}
 
-	width, height := limitedWindowWidthAndHeight(int(C.display_width()), int(C.display_height()))
+	width, height := limitedWindowWidthAndHeight(float64(C.display_width()), float64(C.display_height()))
 
-	w.SetSize(width, height, webview.HintNone)
+	w.SetSize(int(width), int(height), webview.HintNone)
 	w.Navigate(url)
 	bindJSFunctions(w)
 	w.Run()
